Combine test_cache.py with BaseCacheTest to run these tests for all backends
