+++ tests/integration/test_mongodb.py
@@ -5,7 +5,8 @@ from pymongo import MongoClient
 
 from requests_cache.backends import MongoCache, MongoDict, MongoPickleDict, get_valid_kwargs
 from tests.conftest import fail_if_no_connection
-from tests.integration.test_backends import BaseCacheTest, BaseStorageTest
+from tests.integration.base_cache_test import BaseCacheTest
+from tests.integration.base_storage_test import BaseStorageTest
 
 
 @pytest.fixture(scope='module', autouse=True)
