+++ tests/integration/test_redis.py
@@ -3,7 +3,8 @@ from unittest.mock import patch
 
 from requests_cache.backends.redis import RedisCache, RedisDict
 from tests.conftest import fail_if_no_connection
-from tests.integration.test_backends import BaseCacheTest, BaseStorageTest
+from tests.integration.base_cache_test import BaseCacheTest
+from tests.integration.base_storage_test import BaseStorageTest
 
 
 @pytest.fixture(scope='module', autouse=True)
